edit project structure
