+++ lib/app/pages/main/user_profile_page.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:sber_app/app/utils/constants.dart';
+import 'package:sber_app/data/fake_data.dart';
 import 'package:sber_app/app/widgets/appbar.dart';
-import 'package:sber_app/app/widgets/services/services.dart';
+import 'package:sber_app/app/widgets/services_widgets/services.dart';
+import 'package:sber_app/app/theme/colors.dart';
 
 class UserProfilePage extends StatefulWidget {
   const UserProfilePage({super.key});
@@ -21,27 +22,28 @@ class _UserProfilePageState extends State<UserProfilePage>
   }
 
   @override
+  void dispose() {
+    _tabController.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: AppColors.primary,
-      body: SafeArea(
-        child: NestedScrollView(
-            headerSliverBuilder: (context, value) {
-              return [Appbar(tabController: _tabController, user: user)];
-            },
-            body: TabBarView(controller: _tabController, children: [
-              ListView(
-                  physics: const NeverScrollableScrollPhysics(),
-                  children: [
-                    const SizedBox(height: 30),
-                    Services(chapter: enabled, services: services)
-                  ]
-                ),
-              const Center(child: CircularProgressIndicator())
-            ]
-          )
-        )
-      )
-    );
+        backgroundColor: AppColors.primary,
+        body: SafeArea(
+            child: NestedScrollView(
+                headerSliverBuilder: (context, value) {
+                  return [Appbar(tabController: _tabController, user: user)];
+                },
+                body: TabBarView(controller: _tabController, children: [
+                  ListView(
+                      physics: const NeverScrollableScrollPhysics(),
+                      children: [
+                        const SizedBox(height: 30),
+                        Services(chapter: enabled, services: services)
+                      ]),
+                  const Center(child: CircularProgressIndicator())
+                ]))));
   }
 }
