+++ lib/app/sber_app.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:sber_app/app/pages/main/user_profile_page.dart';
-import 'package:sber_app/app/theme.dart';
+import 'package:sber_app/app/theme/theme.dart';
 
 class SberApp extends StatelessWidget {
   const SberApp({super.key});
