+++ lib/app/theme/colors.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class AppColors {
+  static const Color primary = Colors.white;
+  static const Color textPrimary = Colors.black;
+  static final Color textSecondary = Colors.black.withOpacity(0.55);
+  static const Color shadow = Color(0x7A1D1D25);
+  static const Color tabBrand = Color(0xFF068441);
+}
