+++ lib/app/theme/images.dart
@@ -1,0 +1,5 @@
+abstract class Images {
+  static const String userProfileImage = "assets/images/photo.png";
+  static const String sberPrimeImage = "assets/images/sber_prime.png";
+  static const String transfersImage = "assets/images/transfers.png";
+}
