+++ lib/app/theme/strings.dart
@@ -1,0 +1,21 @@
+abstract class Strings {
+  static const String profile = "Профиль";
+  static const String settings = "Настройки";
+  static const String userName = "Екатерина";
+  static const String enabledTitle = "У вас подключено";
+  static const String enabledDescription =
+      "Подписки, автоплатежи и сервисы, на которые вы подписаны";
+  static const String tariffsAndLimitsTitle = "Тарифы и лимиты";
+  static const String tariffsAndLimitsDescription =
+      "Для операций в Сбербанк Онлайн";
+  static const String interestsTitle = "Интересы";
+  static const String interestsDescription =
+      "Мы подбираем истории и предложения по темам, которые вам нравятся";
+  static const String sberPrimeTitle = "СберПрайм";
+  static const String sberPrimePayment = "Платеж 9 июля";
+  static const String sberPrimeCost = "199 рублей в месяц";
+  static const String transfersTitle = "Переводы";
+  static const String transfersPayment = "Автопродление 21 августа";
+  static const String transfersCost = "199 рублей в месяц";
+  
+}
